# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/plans/{planId}/validate": {
      "get": {
        "operationId": "PlansController_validatePlan",
        "summary": "Validate plan availability by id",
        "parameters": [
          {
            "name": "planId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "plans"
        ]
      }
    },
...
    "/v1/doctor-dashboard/patients": {
      "get": {
        "operationId": "DoctorDashboardController_listPatients",
        "summary": "List linked patients",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          },
          "400": {
            "description": "Validation error or invalid input"
          },
          "401": {
            "description": "Missing or invalid JWT"
          },
          "403": {
            "description": "Insufficient role permissions"
          },
          "404": {
            "description": "Resource not found"
          }
        },
        "tags": [
          "doctor-dashboard"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      },
      "post": {
        "operationId": "DoctorDashboardController_addPatient",
        "summary": "Link patient to doctor",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddPatientDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          },
          "400": {
            "description": "Validation error or invalid input"
          },
          "401": {
            "description": "Missing or invalid JWT"
          },
          "403": {
            "description": "Insufficient role permissions"
          },
          "404": {
            "description": "Resource not found"
          }
        },
        "tags": [
          "doctor-dashboard"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/v1/doctor-dashboard/appointments/today": {
      "get": {
        "operationId": "DoctorDashboardController_getTodayAppointments",
        "summary": "Get today's appointments",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          },
          "400": {
            "description": "Validation error or invalid input"
          },
          "401": {
            "description": "Missing or invalid JWT"
          },
          "403": {
            "description": "Insufficient role permissions"
          },
          "404": {
            "description": "Resource not found"
          }
        },
        "tags": [
          "doctor-dashboard"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/v1/doctor-dashboard/online-status": {
      "patch": {
        "operationId": "DoctorDashboardController_updateOnlineStatus",
        "summary": "Toggle doctor online status",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateOnlineStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          },
          "400": {
            "description": "Validation error or invalid input"
          },
          "401": {
            "description": "Missing or invalid JWT"
          },
          "403": {
            "description": "Insufficient role permissions"
          },
          "404": {
            "description": "Resource not found"
          }
        },
        "tags": [
          "doctor-dashboard"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    }
  },
  "info": {
    "title": "EstetoOne API",
    "description": "REST API for EstetoOne (telehealth platform). All routes are prefixed with `/v1` unless noted. Authenticate with Bearer JWT; refresh tokens are issued as HttpOnly `refresh_token` cookies.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3000/v1",
      "description": "Local development"
    }
  ],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "in": "header"
      },
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "refresh_token",
        "description": "HttpOnly refresh token cookie (login/register/refresh)"
      },
      "stripe-signature": {
        "type": "apiKey",
        "in": "header",
        "name": "stripe-signature"
      }
    },
    "schemas": {
...
      "PlanResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "object",
            "nullable": true
          },
          "priceMonthly": {
            "type": "number"
          },
          "features": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "id",
          "type",
          "name",
          "priceMonthly",
          "features"
        ]
      },
...
      "AddPatientDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateOnlineStatusDto": {
        "type": "object",
        "properties": {}
      }
    }
  }
}